+++ test/source/test_sphere_n.cpp
@@ -114,8 +114,8 @@ TEST_CASE("Test Sphere3 reseed functionality") {
     }
 
     // Should be identical
-    for (int i = 0; i < 3; ++i) {
-        for (int j = 0; j < 4; ++j) {
+    for (size_t i = 0; i < 3; ++i) {
+        for (size_t j = 0; j < 4; ++j) {
             CHECK(seq1[i][j] == doctest::Approx(seq2[i][j]).epsilon(1e-10));
         }
     }
@@ -129,8 +129,8 @@ TEST_CASE("Test Sphere3 reseed functionality") {
 
     // Should be different from seed 0
     bool different = false;
-    for (int i = 0; i < 3 && !different; ++i) {
-        for (int j = 0; j < 4; ++j) {
+    for (size_t i = 0; i < 3 && !different; ++i) {
+        for (size_t j = 0; j < 4; ++j) {
             if (std::abs(seq1[i][j] - seq3[i][j]) > 1e-10) {
                 different = true;
                 break;
@@ -187,8 +187,8 @@ TEST_CASE("Test SphereN reseed functionality") {
     }
 
     // Should be identical
-    for (int i = 0; i < 3; ++i) {
-        for (int j = 0; j < 4; ++j) {
+    for (size_t i = 0; i < 3; ++i) {
+        for (size_t j = 0; j < 4; ++j) {
             CHECK(seq1[i][j] == doctest::Approx(seq2[i][j]).epsilon(1e-10));
         }
     }
@@ -255,7 +255,7 @@ TEST_CASE("Sphere3 thread safety") {
                 local_results.push_back(sgen.pop());
             }
             std::lock_guard<std::mutex> lock(mtx);
-            results[i] = std::move(local_results);
+            results[static_cast<size_t>(i)] = std::move(local_results);
         });
     }
     
@@ -293,7 +293,7 @@ TEST_CASE("SphereN thread safety") {
                 local_results.push_back(sgen.pop());
             }
             std::lock_guard<std::mutex> lock(mtx);
-            results[i] = std::move(local_results);
+            results[static_cast<size_t>(i)] = std::move(local_results);
         });
     }
     
@@ -331,7 +331,7 @@ TEST_CASE("SphereWrapper thread safety") {
                 local_results.push_back(sgen.pop());
             }
             std::lock_guard<std::mutex> lock(mtx);
-            results[i] = std::move(local_results);
+            results[static_cast<size_t>(i)] = std::move(local_results);
         });
     }
     
@@ -372,9 +372,9 @@ TEST_CASE("Concurrent reseed thread safety for sphere classes") {
                 if (j % 5 == 0) {
                     // Occasionally reseed
                     if (i % 2 == 0) {
-                        sgen3.reseed(i * 10 + j);
+                        sgen3.reseed(static_cast<size_t>(i * 10 + j));
                     } else {
-                        sgenN.reseed(i * 10 + j);
+                        sgenN.reseed(static_cast<size_t>(i * 10 + j));
                     }
                     reseed_count++;
                 } else {
